+++ cypress/e2e/utm.cy.js
@@ -28,7 +28,7 @@ describe("UTM tests ", () => {
     });
     it("Checks if default utm tags works", () => {
         hp.haltAndClearStorage(() => {
-            initMulti("YOUR_APP_KEY", "?utm_source=hehe&utm_medium=hehe1&utm_campaign=hehe2&utm_term=hehe3&utm_content=hehe4", undefined);
+            initMulti("YOUR_APP_KEY", "utm_source=hehe&utm_medium=hehe1&utm_campaign=hehe2&utm_term=hehe3&utm_content=hehe4", undefined);
             cy.fetch_local_request_queue().then((rq) => {
                 cy.log(rq);
                 const custom = JSON.parse(rq[0].user_details).custom;
@@ -50,7 +50,7 @@ describe("UTM tests ", () => {
     });
     it("Checks if custom utm tags works", () => {
         hp.haltAndClearStorage(() => {
-            initMulti("YOUR_APP_KEY", "?utm_aa=hehe&utm_bb=hoho", { aa: true, bb: true });
+            initMulti("YOUR_APP_KEY", "utm_aa=hehe&utm_bb=hoho", { aa: true, bb: true });
             cy.fetch_local_request_queue().then((rq) => {
                 cy.log(rq);
                 const custom = JSON.parse(rq[0].user_details).custom;
@@ -69,13 +69,13 @@ describe("UTM tests ", () => {
             initMulti("Countly_4", "?utm_source=hehe4", { ss: true });
 
             // utm object not provided with default query
-            initMulti("Countly_3", "?utm_source=hehe3", undefined);
+            initMulti("Countly_3", "utm_source=hehe3", undefined);
 
             // utm object not provided with inappropriate query
             initMulti("Countly_5", "?utm_ss=hehe5", undefined);
 
             // default (original) init with no custom tags and default query
-            initMulti("YOUR_APP_KEY", "?utm_source=hehe", undefined);
+            initMulti("YOUR_APP_KEY", "utm_source=hehe", undefined);
 
             // check original
             cy.fetch_local_request_queue().then((rq) => {
@@ -110,13 +110,13 @@ describe("UTM tests ", () => {
             initMulti("YOUR_APP_KEY", "?utm_source=hehe", undefined);
 
             // utm object not provided with full +  weird query
-            initMulti("Countly_multi_1", "?utm_source=hehe&utm_medium=hehe1&utm_campaign=hehe2&utm_term=hehe3&utm_content=hehe4&fdsjhflkjhsdlkfjhsdlkjfhksdjhfkj+dsf;jsdlkjflk+=skdjflksjd=fksdfl;sd=sdkfmk&&&", undefined);
+            initMulti("Countly_multi_1", "utm_source=hehe&utm_medium=hehe1&utm_campaign=hehe2&utm_term=hehe3&utm_content=hehe4&fdsjhflkjhsdlkfjhsdlkjfhksdjhfkj+dsf;jsdlkjflk+=skdjflksjd=fksdfl;sd=sdkfmk&&&", undefined);
 
             // utm object given that includes 2 default 1 custom, full plus custom query + gabledeboop
             initMulti("Countly_multi_2", "?utm_source=hehe&utm_medium=hehe1&utm_campaign=hehe2&utm_term=hehe3&utm_content=hehe4&utm_sthelse=hehe5&fdsjhflkjhsdlkfjhsdlkjfhksdjhfkj+dsf;jsdlkjflk+=skdjflksjd=fksdfl;sd=sdkfmk&&&", { source: true, term: true, sthelse: true });
 
             // empty init, garbage query + 1 default
-            initMulti("Countly_multi_3", "?dasdashdjkhaslkjdhsakj=dasmndlask=asdkljska&&utm_source=hehe", undefined);
+            initMulti("Countly_multi_3", "dasdashdjkhaslkjdhsakj=dasmndlask=asdkljska&&utm_source=hehe", undefined);
 
             // full default utm obj + custom 1, full query + 1
             initMulti("Countly_multi_4", "?utm_source=hehe&utm_medium=hehe1&utm_campaign=hehe2&utm_term=hehe3&utm_content=hehe4&utm_next=hehe5", { source: true, medium: true, campaign: true, term: true, content: true, next: true });
@@ -171,7 +171,7 @@ describe("UTM tests ", () => {
             initMulti("Countly_multi_next_1", "?utm_sourcer=hehe&utm_mediumr=hehe1&utm_campaignr=hehe2&utm_rterm=hehe3&utm_corntent=hehe4&fdsjhflkjhsdlkfjhsdlkjfhksdjhfkj+dsf;jsdlkjflk+=skdjflksjd=fksdfl;sd=sdkfmk&&&", undefined);
 
             // utm object default, custom query + gabledeboop
-            initMulti("Countly_multi_next_2", "?utm_sourcer=hehe&utm_mediumr=hehe1&utm_campaignr=hehe2&utm_rterm=hehe3&utm_corntent=hehe4&fdsjhflkjhsdlkfjhsdlkjfhksdjhfkj+dsf;jsdlkjflk+=skdjflksjd=fksdfl;sd=sdkfmk&&&", { source: true, medium: true, campaign: true, term: true, content: true });
+            initMulti("Countly_multi_next_2", "utm_sourcer=hehe&utm_mediumr=hehe1&utm_campaignr=hehe2&utm_rterm=hehe3&utm_corntent=hehe4&fdsjhflkjhsdlkfjhsdlkjfhksdjhfkj+dsf;jsdlkjflk+=skdjflksjd=fksdfl;sd=sdkfmk&&&", { source: true, medium: true, campaign: true, term: true, content: true });
 
             // custom utm object, custom query + gabledeboop
             initMulti("Countly_multi_next_3", "?utm_sauce=hehe&utm_pan=hehe2&dasdashdjkhaslkjdhsakj=dasmndlask=asdkljska&&utm_source=hehe", { sauce: true, pan: true });